fix: pr review fixes
