+++ packages/shared/src/lib/core/activities/actions/index.ts
@@ -1,1 +1,2 @@
+export * from './calculateAndAddPersistedBalanceChange'
 export * from './generateActivitiesFromBalanceChanges'
