+++ packages/shared/src/lib/core/activities/types/asset-balance-change.interface.ts
@@ -1,5 +1,5 @@
 export interface IAssetBalanceChange {
-    changedAt: Date
+    changedAt: number
     oldBalance?: number
     newBalance: number
 }
