+++ packages/shared/src/lib/core/activities/utils/generateBalanceChangeActivity.ts
@@ -15,12 +15,12 @@ export function generateBalanceChangeActivity(
     return {
         type: ActivityType.Basic,
         id: generateRandomId(),
-        time: new Date(),
+        time: new Date(balanceChange.changedAt),
         inclusionState: InclusionState.Confirmed,
         containsValue: true,
         isAssetHidden: false,
         direction,
-        action: ActivityAction.BalanceChange,
+        action: balanceChange.oldBalance !== undefined ? ActivityAction.BalanceChange : ActivityAction.InitialBalance,
         isInternal: false,
         storageDeposit: 0,
         subject: undefined,
