+++ packages/shared/src/lib/core/wallet/enums/activity-action.enum.ts
@@ -2,6 +2,7 @@ export enum ActivityAction {
     Send = 'send',
     Burn = 'burn',
     Mint = 'mint',
+    InitialBalance = 'initialBalance',
     BalanceChange = 'balanceChange',
     Unknown = 'unknown',
 }
