+++ packages/shared/src/lib/core/wallet/utils/getActivityTileTitle.ts
@@ -33,6 +33,8 @@ export function getActivityTileTitle(activity: Activity): string {
         return isConfirmed ? 'general.burned' : 'general.burning'
     } else if (action === ActivityAction.BalanceChange) {
         return 'general.balanceChanged'
+    } else if (action === ActivityAction.InitialBalance) {
+        return 'general.initialBalance'
     } else if (action === ActivityAction.Send) {
         if (isInternal) {
             return isConfirmed ? 'general.transfer' : 'general.transferring'
